requirements.txt
asgiref==3.11.0
attrs==25.4.0
cfgv==3.5.0
distlib==0.4.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
django-guardian==3.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
filelock==3.20.3
identify==2.6.15
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Markdown==3.10
mysqlclient==2.2.7
nodeenv==1.10.0
packaging==25.0
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
ruff==0.14.11
ruff-format==0.4.1
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
virtualenv==20.36.1
